userId 解析

+++ star-infrastructure/src/main/java/com/starnft/star/infrastructure/entity/user/AccountUserEntity.java
@@ -61,7 +61,7 @@ public class AccountUserEntity extends BaseEntity implements Serializable {
     /**
      * 是否实名认证
      */
-    private Byte realPersonFlag;
+    private Integer realPersonFlag;
 
     private static final long serialVersionUID = 1L;
 }
